backend/requirements.txt

dj-database-url==0.5.0
Django==3.0.1
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0

